requirements.txt
# Core dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6

# Data processing
pandas>=2.0.0

# Frontend
streamlit>=1.28.0
requests>=2.31.0

# AI/ML
google-generativeai>=0.3.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4

# Utilities
python-dotenv>=1.0.0

# Testing (optional, for development)
pytest>=7.4.0
pytest-cov>=4.1.0
httpx>=0.25.0

# Code quality (optional, for development)
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.0
